Cleaned Formatting
I hate PEP8's insistence on 79 char lines, but this now passes for
99 char lines, aside from a string literal that it wants a space
after a , despite that not being bloody code.

Also rationalized a couple exceptions to do more of what you'd
expect from them, but they still happen in the same places and all.

Still haven't decided how to deal with a trim ending with 0.