Accept (N,0) as valid trim
Trims ending in 0 are now silently turned into the number of frames
reported by the video input, so all math continues to work. It also does
this before the validity check, so that didn't have to change.

I don't actually feel great about that map, to be fair, but only because
the lambda doing a one-line if seems wonky. It's probably just me.