Add doublecheck feature
Adds an option argument to split() that gives a splice of the frames
before and after a given cut point together as a single frame, so you
can step through and ensure things look like you're cutting on the right
frames. This is convenient if you're regularly entering numbers
backwards like a dyslexic person, or if you actually are dyslexic for
realsies (I'm just dumb).