Finally modern MKVToolnix safe
There's a couple bugfixes in here for things like the split checker when a segment starts on frame 0 or ends on the final frame, as well as a fix for '[' or ']' appearing in the audio file name, but those aren't why you're here.

The big one is that this version finally supports mkvtoolnix after the deprecation of --identify-for-mmg. There hasn't been a ton of need to update mkvtoolnix for ages, so I mostly just stuck in place and ignored it, but with multiple issues opened on the tracker it felt like time. Now the flagging of SBR HE-AAC/AAC+ is completely manual. This isn't really likely to ever be an issue for anyone, however, as this feature is only for audio (and presumably video) too low bitrate to be of any value for an encoder. On the plus side, even if you'd been working around the issue with an old mkvmerge.exe you socked away, this version should be faster thanks to not even invoking an external program for that check!